hato-bot用おみくじセットアップをplugins/hato_mikujiに分離 #876
